density change, still a problem with fluidGroup

+++ engine/types.h
@@ -6,9 +6,9 @@
 using Step = uint64_t;
 using Temperature = uint32_t; // degrees kelvin.
 using CollisionVolume = uint32_t; // 20 liters, onehundredth of a block.
-using Volume = uint32_t; // 10 cubic centimeters.
+using Volume = uint32_t; // cubic centimeters.
 using Mass = uint32_t; // grams.
-using Density = uint32_t; // grams per 10 cubic centimeters.
+using Density = float; // grams per cubic centimeter.
 using Force = uint32_t;
 using Percent = int32_t;
 using Facing = uint8_t;
